fix the fix
PS:
ints are: 1,2,3,4... or 0,-1,-2...
numbers with comma (or dot) are real.
We call them float. Or double..

PPS:
I'm maths graduated :D

+++ src/tabpreviewpopup.cpp
@@ -43,7 +43,7 @@
 // static
 static const int borderRadius = 5;
 static const int borderSpacing = 2;
-static const int transparency = 0.90;
+static const double transparency = 0.90;
 
 
 TabPreviewPopup::TabPreviewPopup(WebTab* tab, QWidget* parent)
